Replace long with ptrdiff_t
Make the CI pipeline happy.

+++ src/dcd/common/messages.d
@@ -242,7 +242,7 @@ bool sendRequest(Socket socket, AutocompleteRequest request)
 AutocompleteResponse getResponse(Socket socket)
 {
 	ubyte[1024 * 24] buffer;
-	long bytesReceived = 0;
+	ptrdiff_t bytesReceived = 0;
 	auto unpacker = StreamingUnpacker([]);
 
 	do {
